+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add MetaLogger.Formatter protocol.
+- `log\3` accepts a list as the payload. Each element of the list will be logged separately.
 
 ## [1.2.0] - 2021-04-23
 ### Added